Added casts to avoid compiler warnings
Implicit type conversion warnings were emitted when using
'shader_info::set_specialization_constant' and
'command_buffer_t::draw_indexed'

+++ include/avk/shader_info.hpp
@@ -61,7 +61,7 @@ namespace avk
 			// copy data:
 			memcpy(data.data() + currentSize, &aValue, insertSize);
 			// make entry:
-			entries.emplace_back(aConstantId, currentSize, insertSize);
+			entries.emplace_back(aConstantId, static_cast<uint32_t>(currentSize), insertSize);
 
 			return *this;
 		}
